Refactor: Split calculator into modules
Move shared types to src/lib/calculator-types.ts and extract
graph construction + SCC logic into graph-builder.ts and flow
calculation + SCC solving into flow-solver.ts. Update calculator.ts
to import the new modules and keep behavior unchanged.
